+++ yarte_helpers/src/lib.rs
@@ -1,3 +1,4 @@
 pub use std::fmt::Error;
 pub type Result<I> = ::std::result::Result<I, Error>;
+
 pub mod helpers;